pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[project]
name = "bix"
version = "0.1"
description = "New BIL for Lowell Instruments"
dependencies = [
    "bleak",
    "PyQt6",
    "setproctitle",
    "pyqtgraph",
    "plotly",
    "toml",
    "https://github.com/LowellInstruments/ble.git",
    "https://github.com/LowellInstruments/lix.git",
]